Fix a few incorrect source locations in logs
Motivation:

swift-log 1.3.0 introduced a new concepct: the source of a log. This
identifies the creator of the emitted log, typically the module or
system that emitted the log.

The default value is currently (meant to be) the module name. This is
found by parsing the '#file' to find the directory the file is
contained in. However, this isn't always the module.

Modifications:

- Capture the log source in tests
- Validate the source in test tear down
- Fix up incorrect sources

Result:

- gRPC more uniformly emits a useful log source