Add missing calls to 'close()' in examples
Motivation:

Some of the examples are bad examples: they fail to close connections
which they create even though they close the event loop that they run
on. As such the connection manager has no idea that it no longer needs
to manage a connection which can result in a leaking promise.

Modifications:

- add a defer block to `close()` the connection after we're done with it
  to the echo and hello world examples
- remove the bootstrapping of the logging system (this is no longer
  required in order to quieten the logs since the user must provide the
  logger)

Result:

- HelloWorld example does not leak a promise on shutdown
- Resolves #963