Update the connection ID on transient failures
Motivation:

The logger for the connection mananger includes some metadata for the
connection ID, and the connection mananger is willing to append that
metadata to any logger provided to it.

The ID is a UUID suffixed with a monotonically increasing
number. The suffix is incremented after the connection has become idle
so that new channels have a unique but related ID. However, if a
connection is dropped the connection manager will attempt to
re-establish it yet the suffix will not change. This is quite
misleading in logs since the underlying channel will have changed.

Modifications:

- Increment the connection number when transitioning to transient
  failure

Result:

- The connection ID is incremented when the underlying channel will
  change