Basic Async Unary implementation of QPS benchmark
Motivation:

The main grpc repo contains a benchmarking suite which can be implemented in any language.
This is an implementation of the async unary test and the framework surrounding it.

Notes:
  * Benchmark so far only run on OSX, not linux
  * Benchmark is not optimised - there are currently allocations within the benchmark which
    could be removed.

Modifications:

Add framework for QPS style testing.
Add implementation of Async Unary benchmark (without TLS)
Add brief README describing how to run the benchmark

Result:

It's possible to run a benchmark of gRPC-swift which is in some way comparible to other
languages.