Provide a "call" object for clients
Motivation:

To invoke an RPC you must know the type of RPC you are calling (unary,
client streaming, etc.), its path (e.g. "/echo.Echo/Get"), and the
request and response type it expects as well as some other ancillary
objects to provide the transport for the call. At the moment the type of
the RPC is encoded as the object encapsulating it, e.g. we have a
`UnaryCall` for unary calls and `ClientStreamingCall` for client
streaming calls, and so on. We need something below this: a `Call`
object which can be told what type of call it is. This requirement is
particularly important for interceptors which may retry RPCs, since
within an interceptor we are only aware of the type of a call rather
than the type of the call object.

An additional advantange of this is that it enables us to greatly
simplify the requirements of `GRPCChannel`; instead of implementors
being required to have methods which create a call of each type for each
payload type (`SwiftProtobuf.Message` and `GRPCPayload`) - 8 methods in
total - they may instead have a requirement of making a `Call` for each
of the two payload types (2 methods). The higher level types can be
implemented as default extensions on `GRPCChannel` although that is not
done in this change (to keep it a resonable size).

Modifications:

- Add a `Call` object
- Add a `ClientTransportFactory` for making and configure `ClientTransport`
- Add call making functions to `GRPCChannel` and each implementation
- Add the 'type' and 'path' of the RPC to the interceptor context

Result:

Users can create a `Call` from an interceptor pipeline.