Avoid allocations when accessing interceptor head/tail (#1039)
Motivation:

In some cases array will allocate when accessing first/last (SR-11262),
this happens to occur whenever we access the head or tail of the
interceptor list. This is rather unfortunate as this happens for every
single request/response part we see. Moreover, we know that if the
RPC is still active we will always have a head and a tail.

Modifications:

- Add an 'InterceptorContextList' smaller wrapper around the first and
  last (non-nil) elements and any others.
- Use the context list in the client and server interceptor pipelines.

Result:

- Approximately 3% improvement in QPS benchmarks