Add HTTP2 to gRPC server state machine
Motivation:

We use the HTTP 2-to-1 codecs on the server to simplify the
implementation. However this comes at a performance cost. We removed
the HTTP/1 types from surface API in #1017; this change introduces a
state machine for a translating between HTTP2 frame payloads and gRPC
server request parts. Future PRs will build on top of this to replace
both the 'HTTP1ToGRPCServerCodec' and the 'GRPCServerRequestRoutingHandler'.

Modifications:

Add an 'HTTP2ToRawGRPCStateMachine' and tests.

Result:

We have a state machine in place on which we can build a channel
handler.