Ensure we're on the event loop when sending streaming responses
Motivation:

We expect to send streamed responses from the event loop on the server.
Previously this was enforced by accident as responses were sent by
appending a callback to write the response to a `Channel` future.
However, since the implementation changed we lost this enforcement and
it's possible for users to run into assertion failures if they send
responses from off of the event loop.

Modifications:

Ensure `sendResponse` and `sendResponses` are called on the event loop
in the server streaming call contexts.

Result:

Responses can safely be sent from a streaming response context from off
of the event loop.