Add a 'closeFuture' to the server context
Motivation:

Freeing up user-allocated resources for RPCs is difficult to do
correctly: there isn't a single obvious way to do it for all types, and
in some cases it's impossible since the handler doesn't always know when
the RPC has been terminated. That makes it nigh on impossible for users
to correctly manage resources.

Modifications:

- Add a 'closeFuture' to the server context allowing users to register a
  callbacks to be executed when the RPC completes.
- In existing context implementations this is just the stream channels
  'closeFuture'
- Tests for each RPC style: when the RPC completes successfully, when
  the RPC handler future fails, and when the connection is killed
  mid-rpc

Results:

- It's easier for users to tear down resources when the RPC has
  completed
- Resolves #1145