Add the underlying connection pool
Motivation:

To provide a client (for #1034) which can make RPCs across a number of
connections we need a connection pool which spins up connections and
tears them down as and when necessary and tracks available HTTP/2 streams.

Modifications:

Add a connection pool which callers can ask for a multiplexer "now" or
ask for one in the future if one isn't available.

Result:

We have a connection pool we can build a pooled client on top of.