Add an http/2 connection delegate
Motivation:

As part of #1034 we need some way to notify a connection pool about the
number of active streams on a connection and when we receive a change to
SETTINGS_MAX_CONCURRENT_STREAMS.

Modifications:

Add an internal `HTTP2ConnectionDelegate` protocol and allow it to be
plugged in to the connectivity state monitor. Wire it up to the idle
handler where we recieve notifications of the relevant events.

Result:

The connection pool can be notified when an http/2 stream is closed or
when SETTINGS_MAX_CONCURRENT_STREAMS changes.