Extract the connection monitor from the connection manager (#1173)
Motivation:

The connection monitor allows users to register a delegate for changes
in connectivity state as well as access the current state. Because the
user code may call out it is executed on a `DispatchQueue` (rather than
the connection's `EventLoop`).

However, there are cases within gRPC where state change notifications
on the event loop are very useful (such as in a pool of connection
managers all running on the same event loop).

Moreover, it's not necessary for the connection manager to hold the
connection monitor, it just needs a way to inform it about any changes
so that it can call any user provided delegates.

Modifications:

- Extract the connectivity monitor from the connection manager
- Allow on-event-loop callbacks to be passed to the connection manager
  for connection state changes and for certain HTTP/2 events (stream
  closed and receiving max concurrent streams settings). While not
  strictly necessary for this change they will be helpful down the line.
- Update a bunch of tests.

Result:

The connection manager no longer holds a connection monitor and may
execute state change callbacks on the event loop.