In-source fuzzing support (#1177)
Motivation:

gRPC Swift was recently added to OSS Fuzz. Doing so required patching
gRPC to make the server pipeline configurator public and a small driver
program to setup and run a server on an embedded channel. We should make
supporting oss-fuzz a little nicer!

Modifications:

- Add a FuzzTesting sub package with a 'ServerFuzzer' target
- Symlink in the echo model and implementation
- Add an underscore-public API addition to GRPC to configure an
  `EmbeddedChannel` for fuzzing

Result:

Better support for oss-fuzz