Provide 'default' static functions for making client/server configuration
Motivation:

The `ClientConnection` and `Server` configuration objects are reasonably
large and contain many deafult initialized fields. Whenever we add a
field we need to add a new initializer and deprecate the old one in
order to not break API. Providing a static 'default' function which only
accepts the required properties allows us to add defaulted values without
doing the deprecate and replace.

Modifications:

- Add `default` methods to `ClientConnection.Configuration` and
  `Server.Configuration`
- Deprecate `init`s.

Result:

Easier to add new fields in the future.