Merge pull request from GHSA-4rhq-vq24-88gw
Motivation:

HTTP2ToRawGRPCStateMachine has a parse loop that will attempt to read
messages out of the received buffer. This parse loop is sometimes
recursive: if there are bytes still in the message buffer then we'll
recursively attempt to parse more data.

Swift doesn't do tail-recursion in this pass, so we'll add a frame here.
The minimum message size is 5 bytes: a compression byte and then 4
length bytes. It's therefore possible to produce a very deep call stack
in a single read by sending a long sequence of empty messages.

In fuzzing this call stack can get so deep that it blows the stack
entirely. This is harder to do from the actual network, but still
practical. Fortunately, real applications are usually protected by the
fact that an empty proto message is highly unlikely to be valid for the
specific endpoint that a peer hits.

Still, there's no reason to run the risk, and it's good to be
fuzzing-friendly.

Modifications:

Replace the recursion with looping.

Result:

No longer blow the stack when the peer produces a prodigious set of tiny
messages.

Co-authored-by: Cory Benfield <lukasa@apple.com>