Implement `_preconditionSafeToSyncShutdown` extension point
Motivation:

NIO recently added a `_preconditionSafeToSyncShutdown` extension point
to `EventLoopGroup`. In our tests we have a custom `EventLoopGroup`
called `EmbeddedEventLoopGroup`. When testing against the `main` branch of
swift-nio we hit this precondition (because the default implementation
preconditions that we aren't on any of the event loops in the group, but
we are always on an `EmbeddedEventLoop`).

Modifications:

- Implement `_preconditionSafeToSyncShutdown` for `EmbeddedEventLoopGroup`

Result:

Tests shouldn't crash against NIOs `main` branch.