Use sync options to get the http2 stream ID
Motivation:

NIO added API for getting channel options synchronously. NIO HTTP/2
added support for this in 1.17.0. When configuring a server we configure
our new stream channels to get the stream ID to attach to the loggers.
We can now do this synchronously and save some allocations.

Modifications:

- Bump minimum H2 version to 1.17.0
- Use sync options to get the stream ID and sync pipeline operations to
  add the handler.
- Modify alloc-limits.sh so format better aligns with new CI yaml

Result:

Fewer allocations.