Fix watchOS availability guard for Network.framework TLS
Motivation:

In #1221 we added Network.framework as a provider for TLS. However in a
few places we incorrectly set the availability for watchOS as 5.0. It
should be 6.0 (as required by NIOTS).

If the watchOS deployment target was set to 5.0 then the build would
result in a number of errors (as the availability requirements for using
various NIOTS types would not be sufficient).

Modifications:

- Raise the availability version for various APIs which use Network.framework

Result:

No build errors for types which aren't availabile on watchOS 5.0.