New configuration parameter for HTTP2 maxFrameSize.
Motivation:

In gRPC calls that process large payloads, the default settings of the
HTTP2 maxFrameSize (16384) within swift-nio-http2 results in significant
performance impact (for large payloads this performance impact has been
demonstrated to result in timeouts where RPCs fail to complete within
their assigned deadline).

Allowing programmatic configuration of the HTTP2 maxFrameSize when the
server is built allows the parameter to be optimised for the expected
payload sizes.

Modifications:

Added new gRPC configuration parameter for the HTTP2 parameter
maxFrameSize (which can be configured in the Server.Builder using the
new method `withHTTPMaxFrameSize()`).

Result:

The Server.Builder can be used to configure the HTTP2 maxFrameSize
parameter as:

Server.insecure(group: serverGroup).withHTTPMaxFrameSize(1024 * 1024)...